requirements.txt
numpy
datasets
pyyaml
psycopg2-binary
